Updated Piece interface to include Player owner.
