+++ libGDX/core/src/com/mygdx/game/movement/pieces/Hopper.java
@@ -1,25 +1,27 @@
 package com.mygdx.game.movement.pieces;
 
+import com.mygdx.game.movement.Board;
 import com.mygdx.game.movement.CoOrdinatePair;
+import com.mygdx.game.players.Player;
 
 import java.util.HashSet;
 import java.util.Set;
 
 public class Hopper implements Piece {
 
-  private final boolean isPlayerPiece;
 
-  public Hopper(boolean isPlayerPiece) {
-    this.isPlayerPiece = isPlayerPiece;
+  private Player owner;
+
+  private Board board;
+
+  public Hopper(Player owner) {
+    this.owner = owner;
   }
 
-  public Hopper() {
-    this(true);
-  }
 
   @Override
-  public boolean isPlayerPiece() {
-    return isPlayerPiece;
+  public boolean isPlayerPiece(Player player) {
+    return owner.equals(player);
   }
 
   @Override
@@ -37,5 +39,20 @@ public class Hopper implements Piece {
     moveSet.add(origin.up().up());
     moveSet.add(origin.up().up().up());
     return moveSet;
+  }
+
+  @Override
+  public int value() {
+    return 5;
+  }
+
+  @Override
+  public Player getOwner() {
+    return owner;
+  }
+
+  @Override
+  public Board getBoard() {
+    return board;
   }
 }
