+++ libGDX/core/src/com/mygdx/game/round/Move.java
@@ -1,5 +1,7 @@
 package com.mygdx.game.round;
 
+import com.mygdx.game.players.Player;
+
 @FunctionalInterface
 public interface Move {
   public void execute();
